+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - CI-108: Add back and update sending tagging data to GA 360.
 - CI-150 - Updated featured cards on Invest home page to get all details from cms
 - NO TICKET - Gave featured cards on Invest home page unique ids
+- [CI-126]: Redirect setup pages to new international setup pages. Delete the old pages.
 
 ## [2019.05.16](https://github.com/uktrade/invest-ui/releases/tag/2019.05.16)
 [Full Changelog](https://github.com/uktrade/invest-ui/compare/2019.05.09...2019.05.16)